Update Alex-withcolour_sensor.ino
I have added some variables and map functions to convert the values read by the sensor into color codes. Before implementing this map function we need to calibrate the color sensor ie: we need to find the lower and higher values read by the color sensor. If the serial.print give errors while doing in RPI, we need to use the same algorithm but through our laptop without using RPI and use the calibrated values in the actual code.